requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
python-dotenv==1.0.1
SQLAlchemy==2.0.26
SQLAlchemy-Utils==0.41.1
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==2.3.3
